Revert "Making --livereload and --devicesync synonyms."
Because of the fact that NOPT parses '--device' as '--devicesync', I'm removing '--devicesync' as a possible option.
NOPT's behavior is probably the result of some weird bug in their code.

Also, we don't need the '--devicesync' flag here anymore as this plugin won't be used
from TACO anymore because I've changed the way LiveReload will be included in TACO (from plugin architecture to package-based architecture):
https://github.com/Microsoft/TACO/pull/174

This reverts commit 9fc9038d8f084f263cdf2aa5142849fd32681cd3.